fix(General configuration): Implementation
- Implemented react hook form for group list
